requirements.txt
snowflake-connector-python
mistralai
python-dotenv
sentence-transformers
transformers
pdfminer.six
requests
streamlit
